+++ lib/gui/app/components/settings/settings.tsx
@@ -17,11 +17,12 @@
 import GithubSvg from '@fortawesome/fontawesome-free/svgs/brands/github.svg';
 import * as _ from 'lodash';
 import * as React from 'react';
-import { Box, Checkbox, Flex, Txt } from 'rendition';
+import { Box, Checkbox, Divider, Flex, Txt } from 'rendition';
 
 import { version, packageType } from '../../../../../package.json';
 import * as settings from '../../models/settings';
 import * as analytics from '../../modules/analytics';
+import { open as openInternal } from '../../os/open-internal-remote/services/open-internal';
 import { open as openExternal } from '../../os/open-external/services/open-external';
 import { Modal } from '../../styled-components';
 import * as i18next from 'i18next';
@@ -37,22 +38,19 @@ async function getSettingsList(): Promise<Setting[]> {
 		{
 			name: 'verify',
 			label: i18next.t('settings.verify'),
-		},
-		{
-			name: 'errorReporting',
-			label: i18next.t('settings.errorReporting'),
+			tooltip: i18next.t('settings.verifyDesc'),
 		},
 		{
 			name: 'autoBlockmapping',
 			label: i18next.t('settings.trimExtPartitions'),
+			tooltip: i18next.t('settings.trimExtPartitions'),
+		},
+		{
+			name: 'decompressFirst',
+			label: i18next.t('settings.decompressFirst'),
+			tooltip: i18next.t('settings.decompressFirst'),
 		},
 	];
-	if (['appimage', 'nsis', 'dmg'].includes(packageType)) {
-		list.push({
-			name: 'updatesEnabled',
-			label: i18next.t('settings.autoUpdate'),
-		});
-	}
 	return list;
 }
 
@@ -105,19 +103,27 @@ export function SettingsModal({ toggleModal }: SettingsModalProps) {
 		<Modal
 			titleElement={
 				<Txt fontSize={24} mb={24}>
-					{i18next.t('settings.settings')}
+					<u>
+						{i18next.t('settings.settings')}
+					</u>
 				</Txt>
 			}
 			done={() => toggleModal(false)}
 		>
+			<Txt fontSize={24} mb={14}>
+				<u>
+					{i18next.t('settings.settings')}
+				</u>
+			</Txt>
 			<Flex flexDirection="column">
 				{settingsList.map((setting: Setting, i: number) => {
 					return (
-						<Flex key={setting.name} mb={14}>
+						<Flex key={setting.name} mt={14} mb={14}>
 							<Checkbox
 								toggle
 								tabIndex={6 + i}
 								label={setting.label}
+								tooltip={setting.tooltip}
 								checked={currentSettings[setting.name]}
 								onChange={() => toggleSetting(setting.name)}
 							/>
@@ -134,8 +140,9 @@ export function SettingsModal({ toggleModal }: SettingsModalProps) {
 						)}
 					</Flex>
 				)}
+				<Divider color="#00aeef" />
 				<Flex
-					mt={18}
+					mt={14}
 					alignItems="center"
 					color="#00aeef"
 					title="View Changelog"
@@ -145,7 +152,7 @@ export function SettingsModal({ toggleModal }: SettingsModalProps) {
 						fontSize: 14,
 					}}
 					onClick={() =>
-						openExternal(
+						openInternal(
 							'https://github.com/Alex313031/etcher-ng/blob/master/CHANGELOG.md',
 						)
 					}