+++ lib/gui/app/modules/analytics.ts
@@ -15,7 +15,8 @@
  */
 
 import * as _ from 'lodash';
-import { Client, createClient, createNoopClient } from 'analytics-client';
+import type { Client } from 'analytics-client';
+import { createClient, createNoopClient } from 'analytics-client';
 import * as SentryRenderer from '@sentry/electron/renderer';
 import * as settings from '../models/settings';
 import { store } from '../models/store';
@@ -141,7 +142,7 @@ export const initAnalytics = _.once(() => {
 
 const getCircularReplacer = () => {
 	const seen = new WeakSet();
-	return (key: any, value: any) => {
+	return (_key: any, value: any) => {
 		if (typeof value === 'object' && value !== null) {
 			if (seen.has(value)) {
 				return;
@@ -206,7 +207,7 @@ function reportAnalytics(message: string, data: AnalyticsPayload = {}) {
  * This function sends the debug message to product analytics services.
  */
 export async function logEvent(message: string, data: AnalyticsPayload = {}) {
-	const shouldReportAnalytics = await settings.get('errorReporting');
+	const shouldReportAnalytics = false;
 	if (shouldReportAnalytics) {
 		initAnalytics();
 		reportAnalytics(message, data);
@@ -220,7 +221,7 @@ export async function logEvent(message: string, data: AnalyticsPayload = {}) {
  * This function logs an exception to error reporting services.
  */
 export function logException(error: any) {
-	const shouldReportErrors = settings.getSync('errorReporting');
+	const shouldReportErrors = false;
 	console.error(error);
 	if (shouldReportErrors) {
 		initAnalytics();