+++ lib/gui/app/modules/progress-status.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import * as prettyBytes from 'pretty-bytes';
+import prettyBytes from 'pretty-bytes';
 import * as i18next from 'i18next';
 
 export interface FlashState {
@@ -34,6 +34,8 @@ export function fromFlashState({
 	status: string;
 	position?: string;
 } {
+	console.log(i18next.t('progress.starting'));
+
 	if (type === undefined) {
 		return { status: i18next.t('progress.starting') };
 	} else if (type === 'decompressing') {